A-143: Animation of sprites

+++ atomina-game/GameApp.hpp
@@ -1,11 +1,5 @@
 #include <atomina.hpp>
 #include "GameEnums.hpp"
-#include "state/MainMenuState.hpp"
-#include "state/LobbyState.hpp"
-#include "state/PlayState.hpp"
-#include "OAS/AttrConnection.hpp"
-#include "OAS/SysConnection.hpp"
-#include "OAS/SysTime.hpp"
 
 /*
  * extension of the Game class from atomina-api
@@ -30,7 +24,7 @@ public:
      * real implementation of the game apps run function to house the
      * game loop
      */
-    virtual void update(ATMA::ATMAContext *l_ctx, const long long &l_dt) override;
+    virtual void update(ATMA::ATMAContext *l_ctx, const double &l_dt) override;
 
     /**
      * real implementation of the game apps shut down function
